projects/adk-agents/pyproject.toml
# =============================================================================
# Project Metadata
# =============================================================================
[project]
name = "ds-adk-agents"
version = "0.1.0"
description = "Desert Services ADK-based agentic workflows"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "google-adk>=0.3.0",
    "google-genai>=1.0.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.32.0",
]

# =============================================================================
# Development Dependencies (PEP 735)
# =============================================================================
[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.11",
    "ty>=0.0.11",
    "watchfiles>=1.1.1",
]

# =============================================================================
# Build System
# =============================================================================
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# =============================================================================
# Testing - pytest
# =============================================================================
[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"
testpaths = ["tests"]

# =============================================================================
# Linting & Formatting - ruff
# =============================================================================
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    # Core
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "W",      # pycodestyle warnings
    # Import sorting
    "I",      # isort
    "TID",    # flake8-tidy-imports
    # Code quality
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "PIE",    # flake8-pie
    "SIM",    # flake8-simplify
    "UP",     # pyupgrade
    # Style
    "N",      # pep8-naming
    "Q",      # flake8-quotes
    # Security
    "S",      # flake8-bandit
    # Type checking
    "TCH",    # flake8-type-checking
    # Best practices
    "A",      # flake8-builtins
    "ARG",    # flake8-unused-arguments
    "DTZ",    # flake8-datetimez
    "RET",    # flake8-return
    "RSE",    # flake8-raise
    "SLF",    # flake8-self
    "SLOT",   # flake8-slots
    # Testing
    "PT",     # flake8-pytest-style
    # Debugging (disallow in production)
    "T10",    # flake8-debugger
    "T20",    # flake8-print
    # Advanced linters
    "FLY",    # flynt
    "FURB",   # refurb
    "PERF",   # perflint
    "PGH",    # pygrep-hooks
    "PL",     # pylint
    "RUF",    # ruff-specific
    "TRY",    # tryceratops
]
ignore = [
    "B008",    # function calls in argument defaults (FastAPI pattern)
    "E501",    # line too long (handled by formatter)
    "PLR0913", # too many arguments (agents need flexibility)
    "S101",    # assert used (acceptable in tests)
    "TRY003",  # long exception messages (readability > brevity)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["PLR2004", "S101", "T201"]

[tool.ruff.lint.pylint]
max-args = 10

# =============================================================================
# Type Checking - ty
# =============================================================================
[tool.ty.environment]
python-version = "3.11"
